+++ src/Radical.Windows/Behaviors/FrameworkElementLifecycleNotificationsBehavior.cs
@@ -1,6 +1,6 @@
 ﻿using Radical.ComponentModel.Messaging;
 using Radical.Windows.ComponentModel;
-using Radical.Windows.Presentation.Messaging;
+using Radical.Windows.Messaging;
 using System.Windows;
 
 namespace Radical.Windows.Behaviors
